requirements.txt
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.5
dateutils==0.6.12
et-xmlfile==1.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
kaitaistruct==0.10
openpyxl==3.1.2
outcome==1.3.0.post0
pyasn1==0.5.1
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
selenium==4.18.1
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
tweety-ns==1.0.9.8
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
zstandard==0.22.0
